services/api/pyproject.toml
[project]
name = "lrp-api"
version = "0.1.0"
description = "LRP Scheduling Agent API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosql>=15.0",
    "arq>=0.27.0",
    "fastapi>=0.135.2",
    "nanoid>=2.0.0",
    "psycopg[binary]>=3.3.3",
    "psycopg-pool>=3.2",
    "pydantic>=2.12.5",
    "redis>=5.3.1",
    "sentry-sdk[fastapi]>=2.56.0",
    "uvicorn[standard]>=0.42.0",
    "google-auth>=2.0",
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.0",
    "cryptography>=44.0",
    "yoyo-migrations>=9.0",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "requests>=2.33.1",
    "langfuse>=2.0",
    "litellm>=1.0",
    "anthropic>=0.40",
    "openai>=1.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.28.1",
    "hypothesis>=6.151.10",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"scripts/*" = ["E501"]  # prompt templates have long lines

[tool.ruff.lint.isort]
known-first-party = ["api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
